fix(bigint): mutiplication bug
- when multiplying uint64max by 5, it returned 20 at the 2nd chunk, instead of 4. it was caused by addChunkInt extending chunks.size(), and multiplication loop going 2nd time

+++ src/bigint/bigint.hpp
@@ -7,7 +7,7 @@
 
 void checkBigInt();
 void checkmultip();
-typedef uint32_t chunkInt;
+typedef uint64_t chunkInt;
 
 enum class chunkDisplayMode {
 	decimal,
